requirements.txt
pychef
termcolor
